Add stress tester
This adds a script that produces logs of rit | bs output for a
thousand, ten thousand, a hundred thousand, and a million random ints.
It writes the sorted output to *_out.log and the logged durations of
the runs to *_perf.log in a log folder created for the time the run was
started.